requirements.txt
# # Django==5.0.3
# # django-filter==24.1
# # djangorestframework==3.15.0
# # drf-spectacular==0.27.1

# Django==4.1
# django-filter==22.1
# djangorestframework==3.15.0
# drf-spectacular==0.27.1
# psycopg2-binary==2.9.3
# Gdal == 3.2.2

amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
billiard==3.6.4.0
boto3==1.26.75
botocore==1.29.75
cached-property==1.5.2
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cligj==0.7.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
cycler==0.11.0
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.2.4
Django==4.1
django-allauth==0.51.0
django-cors-headers==3.13.0
django-debug-toolbar==3.4.0
django-filter==22.1
django-storages==1.13.2
djangorestframework==3.13.1
drf-yasg==1.20.0
Fiona==1.8.21
fonttools==4.38.0
GDAL==3.2.2
geopandas==0.11.0
geoserver-rest==2.3.3
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.3.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.3
kombu==5.2.4
MarkupSafe==2.1.1
matplotlib==3.5.2
munch==2.5.0
numpy==1.22.4
oauthlib==3.2.2
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pycparser==2.21
Pygments==2.12.0
PyJWT==2.6.0
pyparsing==3.0.9
pyproj==3.3.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
redis==4.3.3
requests==2.28.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
scipy==1.8.1
seaborn==0.11.1
Shapely==1.8.2
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.9
uWSGI==2.0.21
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xmltodict==0.13.0
drf-spectacular==0.27.1
#fastapi
# fastapi==0.85.2
# sqlalchemy==1.4.42
# alembic==1.8.1
# uvicorn==0.19.0
# databases==0.6.1
# asyncpg==0.27.0
# SecretStorage==3.3.3
# motor==3.1.1
# ejson==0.1.6
# pymongo==4.3.2